[Fixes #68112178] ES indexing of Twitter profile ID's needs to account for long integers
- Also: Spread 11GB Odie doc index across machines

+++ app/models/elastic_tweet.rb
@@ -9,10 +9,10 @@ class ElasticTweet
       _analyzer: { path: "language" },
       properties: {
         language: { type: "string", index: :not_analyzed },
-        twitter_profile_id: { type: 'integer' },
+        twitter_profile_id: { type: 'long' },
         tweet_text: { type: 'string', term_vector: 'with_positions_offsets' },
         published_at: { type: 'date' },
-        id: { type: 'integer', index: :not_analyzed, include_in_all: false } }
+        id: { type: 'long', index: :not_analyzed, include_in_all: false } }
     }
   }
